+++ code/header/block/saturation_block.h
@@ -28,7 +28,7 @@ class SATURATION_BLOCK
         virtual void check();
     private:
         SATURATION_TYPE type;
-        float V1, V2, S1, S2;
+        double V1, V2, S1, S2;
 
         double A, B, C;
         //S=B(V-A)^2/V or
