+++ code/header/model/sg_models/exciter_model/CSEET2.h
@@ -122,14 +122,14 @@ class CSEET2: public EXCITER_MODEL
         PI_BLOCK serial_tuner1_pi;
         LEAD_LAG_BLOCK serial_tuner2;
         // parallel
-        float KP;
+        double KP;
         INTEGRAL_BLOCK parallel_integral;
         DIFFERENTIAL_BLOCK parallel_differential;
         // common
         FIRST_ORDER_BLOCK regulator;
         DIFFERENTIAL_BLOCK feedbacker;
-        float VRmax, VRmin;
-        float KC;
+        double VRmax, VRmin;
+        double KC;
 
 };
