+++ code/header/model/sg_models/exciter_model/PSASPE1.h
@@ -74,7 +74,7 @@ class PSASPE1: public EXCITER_MODEL
         FIRST_ORDER_BLOCK regulator;
         DIFFERENTIAL_BLOCK feedbacker;
         FIRST_ORDER_BLOCK exciter;
-        float VAmax, VAmin;
+        double VAmax, VAmin;
         double Efd0;
 };
