+++ code/header/model/sg_models/turbine_governor_model/IEEEG3.h
@@ -81,11 +81,11 @@ class IEEEG3 : public TURBINE_GOVERNOR_MODEL
 
         FIRST_ORDER_BLOCK governor;
         INTEGRAL_BLOCK servo_motor;
-        float delta, sigma;
+        double delta, sigma;
         DIFFERENTIAL_BLOCK feedbacker;
         LEAD_LAG_BLOCK water_hammer;
-        float TW;
-        float a11, a13, a21, a23;
+        double TW;
+        double a11, a13, a21, a23;
 };
 
 #endif // IEEEG3_H
